Computes the distribution of fitness differences between new and original mutations as well as the frequency with which those are attempted

Computes the distribution of fitness differences between new and original mutations as well as the frequency with which those are attempted

ComputeMutationFitnesses(nuc.model = "JC", base.freqs = rep(0.25, 4),
  nsites = 1, C = 4, Phi = 0.5, q = 4e-07, Ne = 5e+06, alpha = 1.83,
  beta = 0.1, gamma = 0.0003990333, include.stop.codon = TRUE,
  numcode = 1, diploid = TRUE, flee.stop.codon.rate = 0.9999999)

Details

Returns a list; most of the return objects are the same as from ComputeEquilibriumAAFitness() but the new things are codon.mutation.matrix: instantaneous rate matrix for codon mutations codon.relative.rate.matrix: the above, but scaled so that each row sums to 1 delta.fitness.array: 3d array: dimensions are starting codon, optimal aa, and finishing codon; entries are new codon fitness - original codon fitness frequency.array: 3d array: dimensions are starting codon, optimal aa, and finishing codon; entries are frequencies that that mutation is attempted given the optimal aa (column)