Efficient optimization of the SELON model

Efficient optimization of model parameters under the SELON model

SelonLargeOptimize(nuc.data.path, n.partitions = NULL, phy,
  edge.length = "optimize", edge.linked = TRUE, optimal.nuc = "majrule",
  nuc.model = "GTR", diploid = TRUE, verbose = FALSE, n.cores = NULL,
  max.tol = .Machine$double.eps^0.5, max.evals = 1e+06, max.restarts = 10,
  output.by.restart = TRUE, output.restart.filename = "restartResult",
  fasta.rows.to.keep = NULL)

Arguments

nuc.data.path
Provides the path to the directory containing the gene specific fasta files that contains the nucleotide data.

n.partitions
The number of partitions to analyze. The order is based on the Unix order of the fasta files in the directory.

phy
The phylogenetic tree to optimize the model parameters.

edge.length
Indicates whether or not edge lengths should be optimized. By default it is set to "optimize", other option is "fixed", which user-supplied branch lengths.

edge.linked
A logical indicating whether or not edge lengths should be optimized separately for each gene. By default, a single set of each lengths is optimized for all genes.

optimal.nuc
Indicates what type of optimal.nuc should be used. At the moment there is only a single option: "majrule".

nuc.model
Indicates what type nucleotide model to use. There are three options: "JC", "GTR", or "UNREST".

diploid
A logical indicating whether or not the organism is diploid or not.

verbose
Logical indicating whether each iteration be printed to the screen.

n.cores
The number of cores to run the analyses over.

max.tol
Supplies the relative optimization tolerance.

max.evals
Supplies the max number of iterations tried during optimization.

max.restarts
Supplies the number of random restarts.

output.by.restart
Logical indicating whether or not each restart is saved to a file. Default is TRUE.

output.restart.filename
Designates the file name for each random restart.

fasta.rows.to.keep
Indicates which rows to remove in the input fasta files.

Details

This is the exact same implementation as in SelonOptimize(), except here we optimize parameters across each gene separately while keeping the shared parameters, the edge lengths, constant across genes. We then optimize edge lengths while keeping parameters for each gene fixed. This approach is potentially more efficient than simply optimizing all parameters simultaneously, especially if fitting models across 100's of genes.